+++ crates/sickle_ui_scaffold/src/attributes/traits.rs
@@ -1,16 +1,20 @@
+use std::any::Any;
+use std::fmt::Debug;
+
 use bevy::prelude::*;
+use dyn_clone::DynClone;
 
 use crate::*;
 
 //-------------------------------------------------------------------------------------------------------------------
 
 /// Helper trait for storing custom static attributes.
-pub trait StaticAttributeObject: AnyClone
+pub trait StaticAttributeObject: Any + DynClone + Debug + Send + Sync + 'static
 {
-    /// Convert self to an `AnyClone` reference.
-    fn as_anyclone(&self) -> &dyn AnyClone;
-    /// Convert self to a mutable `AnyClone` reference.
-    fn as_anyclone_mut(&mut self) -> &mut dyn AnyClone;
+    /// Convert self to an `Any` reference.
+    fn as_any(&self) -> &dyn Any;
+    /// Convert self to a mutable `Any` reference.
+    fn as_any_mut(&mut self) -> &mut dyn Any;
 
     /// Applies the attribute to the target entity.
     fn apply(&self, entity: Entity, world: &mut World);
@@ -19,12 +23,12 @@ pub trait StaticAttributeObject: AnyClone
 //-------------------------------------------------------------------------------------------------------------------
 
 /// Helper trait for storing custom responsive attributes.
-pub trait ResponsiveAttributeObject: AnyClone
+pub trait ResponsiveAttributeObject: Any + DynClone + Debug + Send + Sync + 'static
 {
-    /// Convert self to an `AnyClone` reference.
-    fn as_anyclone(&self) -> &dyn AnyClone;
-    /// Convert self to a mutable `AnyClone` reference.
-    fn as_anyclone_mut(&mut self) -> &mut dyn AnyClone;
+    /// Convert self to an `Any` reference.
+    fn as_any(&self) -> &dyn Any;
+    /// Convert self to a mutable `Any` reference.
+    fn as_any_mut(&mut self) -> &mut dyn Any;
 
     /// Applies the attribute to the target entity.
     fn apply(&self, entity: Entity, world: &mut World, state: FluxInteraction);
@@ -33,12 +37,12 @@ pub trait ResponsiveAttributeObject: AnyClone
 //-------------------------------------------------------------------------------------------------------------------
 
 /// Helper trait for storing custom animated attributes.
-pub trait AnimatedAttributeObject: AnyClone
+pub trait AnimatedAttributeObject: Any + DynClone + Debug + Send + Sync + 'static
 {
-    /// Convert self to an `AnyClone` reference.
-    fn as_anyclone(&self) -> &dyn AnyClone;
-    /// Convert self to a mutable `AnyClone` reference.
-    fn as_anyclone_mut(&mut self) -> &mut dyn AnyClone;
+    /// Convert self to an `Any` reference.
+    fn as_any(&self) -> &dyn Any;
+    /// Convert self to a mutable `Any` reference.
+    fn as_any_mut(&mut self) -> &mut dyn Any;
 
     /// Initializes self when begining to enter the idle state.
     fn initialize_enter(&mut self, entity: Entity, world: &World);