+++ crates/sickle_ui_scaffold/src/lib.rs
@@ -1,4 +1,3 @@
-mod any_clone;
 mod attributes;
 mod builder_ext;
 mod flux_interaction;
@@ -7,7 +6,6 @@ mod ui_commands;
 mod ui_style;
 mod ui_utils;
 
-pub use any_clone::*;
 pub use attributes::*;
 pub use builder_ext::*;
 pub use flux_interaction::*;